Add third digital channel to IF0
A third digital channel is added to IF0. Initial build did not meet
timing (with a score of 4), and negative slack needed in paths
internal to the second stage DDC, more specifically within the
FIR part of the biquad IIR structure. The pipeline delays on either
side of the second stage DDC were increased to 66 (from 64) to
yield a design that met timing.