Add ADC core calibration routines
Calibration routines were added to ArtooDaq to reduce artefacts that
are introduced into the sampled data spectrum due to the interleaved
cores within the ADC. These routines assume that the input analog
signal is zero-mean and that it will yield approximately equal power
in the samples produced by each of the cores.

The cores within the ADC each has an independent offset, gain and
phase register. The currently implemented calibration routines tune
only the offset and gain parameters.