Remove dependence on external sync pulse
The bitcode was updated so that a synchronization pulse could be
generated manually. This removes the need for an external pulse to
start the on-FPGA processing.

The simulink file (.slx) and the binary program file (in zipped
form .bof.gz) reflect these changes. Copy gunzip-ed program file
to usual location to load onto the ROACH2. Standard practice for
compiled bitcode should be to commit only zipped versions to reduce
commit sizes.

The system interface library r2daq.py includes the manual sync
pulse generation in the startup. It also uses the bitcode committed
here as the 'latest-build' default.

The rcs_user register for this bitcode version is 0x00000001.