Fix ADC core vs interface calibration misnomer
What used to be called 'core' calibration in code, comments, and
output actually pertained to the interface between the ADC and the
FPGA. Various changes have been made to correctly name what should
be called 'interface' calibration.

The actual ADC quad-core calibration is also now done in the
ArtooDaq._start method (if parameter do_cal==True, which is the
default).

A verbosity parameter was added to ADC quad-core calibration methods
to control the level of output during the calibration.