Enable arbitrary data network configuration
The specification of a data receiver network interface parameter in
ArtooDaq._start has been dropped to allow configuring the ROACH2 10GbE
interfaces with arbitrary source / destination addresses.

The class method make_interface_config_dictionary has been added to
ArtooDaq to instantiate a dictionary that encapsulates the configuration
of an interface associated with one particular digital channel. This
method should be called once for each interface that needs configuration
and the results passed in a list to ArtooDaq.__init__ as the keyword
argument ifcfg. If no such list is given then __init__ will create a
default list.

ArtooDaq._config_channel_net does the actual configuration of the
ROACH2 network interfaces. This method will eventually become available
for changing the network configuration of any channel on a running system.
However, in the current implementation it is advisable not to call this
method outside the ArtooDaq._start method.

A number of helper methods have been added to ease switching various
control signals in the bitcode on and off, ArtooDaq._*_(hi|lo). These
methods should for now only be used during execution of ArtooDaq._start.

ArtooDaq._load_time has been added to hide the details of setting the
correct time at startup.

ArtooDaq._tmp_add_iface_dict and the accompanying structure
ArtooDaq._tmp_iface_dict have been added as a workaround to enable
network flexibility. These will be removed once the network can be changed
arbitrarily on a running system.