Propagate ADC calibration flag(s) to ArtooDaq.__init__
A flag has been added to ArtooDaq._start to switch ADC interface
calibration on / off (latter is useful for debugging to save time) and
the flag for core calibration renamed to make distinction between the
two flags clearer.

The calibration flags are now propagated to the __init__ method so that
the automatic programming and configuration does not need to be bypassed
to skip calibration.